Add files via upload
This is a one player game in which your have a spaceship and you have to kill all the nemies that are approaching towards you.  As soon as they touched your spaceship game gets over. You have to play efficiently because those enemies are very clever. Your score is also be counted on upper left hand corner and try to bet your highest score. I hopes that you have understood the game logics and will enjoy my game.